backend/requirements.txt
fastapi
uvicorn
motor
pydantic
passlib[bcrypt]
python-multipart
python-dotenv
pymongo 